+++ Main/Features/Warg/IWargAttackService.cs
@@ -1,10 +1,10 @@
-using TaleWorlds.MountAndBlade;
+using TAOM.Adapters;
 
 namespace TAOM.Features.Warg;
 
 public interface IWargAttackService
 {
-    int CalculateWargAttackDamage(Agent target, float velocity);
-    void HandleWargTargetHit(Agent attacker, Agent target, sbyte boneId);
-    void WargAttack(Agent warg);
+    int CalculateWargAttackDamage(IAgentAdapter target, float velocity, float armorEffectivenessPercent);
+    void HandleWargTargetHit(IAgentAdapter attacker, IAgentAdapter target, sbyte boneId);
+    void WargAttack(IAgentAdapter warg);
 }